==> 1_make_dataset_preprocess.ipynb <==
### Define helper functions

import urllib.request
import json
import numpy as np
import pandas as pd
import altair
import pprint
import os, subprocess, sys,codecs, locale
import re
import traceback
from collections import namedtuple
import datetime
import pdfplumber
%matplotlib inline

pd.options.display.width = 250
pd.options.display.max_colwidth = 250
pd.options.display.max_rows = 250


# some helper functions to obtain the right output from API
def remove_spaces(input):
    return input.replace(' ', '%20')

def pdf_to_text(pdf_path):
    #sys.setdefaultencoding("utf-8")
    os_encoding = locale.getpreferredencoding()
    args = ["pdftotext.exe",pdf_path, "-"]
    res = subprocess.run(args, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
    output = (res.stdout).decode(os_encoding,'ignore')
    return(output)


# this is the preferred method, pdfplumber works better then pdftotext
def pdf_to_text2(pdf_path):
    with pdfplumber.open(pdf_path) as pdf:
        first_page = pdf.pages[0]
        return first_page.extract_text(x_tolerance=2, y_tolerance=3)


def retrieve_and_save_pdf(id, overwrite=False):
    if not os.path.exists(path + id + '.pdf') or overwrite:
        base = 'https://gegevensmagazijn.tweedekamer.nl/OData/v4/2.0/'
        url = 'Document('+id+')/Resource'
        end = ''
        response = urllib.request.urlopen(remove_spaces(base+url))
        data = response.read()      # a `bytes` object
        f = open(path+id+'.pdf', 'w+b')
        f.write(data)
        f.close()
    else:
        pass

def query_API(skip):
    base = 'https://gegevensmagazijn.tweedekamer.nl/OData/v4/2.0/'
    url = "/Zaak?$filter= Soort eq 'Motie'&$expand=Besluit($expand=Stemming),Document,ZaakActor,Agendapunt&$count=true&$skip="+skip
    end = '&$format=application/json;odata.metadata=full'
    response = urllib.request.urlopen(remove_spaces(base+url))
    data = response.read()      # a `bytes` object
    return(json.loads(data))
#retrieve_and_save_pdf('1b7f8f8c-4579-4103-9f51-9280358f7b8a')
#a = pdf_to_text(path+'3c2dbc59-beb1-42d8-9be4-e09238f20a97.pdf')
#print(a)

### Process API response to dict

def add_API_to_dict(data):
    try:
        for z in data['value']:
            #print('\n\n','nummer',z['Nummer'])
            info[z['Nummer']]={'Titel':z['Titel'],
                              #'Status':z['Status'],
                              'Onderwerp':z['Onderwerp'],
                              'Vergaderjaar':z['Vergaderjaar'],
                              'GestartOp':z['GestartOp']}

            for b in z['Besluit']:
                info[z['Nummer']]['BesluitTekst']=b['BesluitTekst']
                info[z['Nummer']]['StemmingsSoort']=b['StemmingsSoort']
                info[z['Nummer']]['BesluitSoort']=b['BesluitSoort']

                #print('\n','besluit',b['BesluitTekst'])
                if b['BesluitSoort'] and b['BesluitSoort'] in ['Stemmen - aangenomen','Stemmen - verworpen', 'Stemmen - niet aangenomen']:        
                    for s in b['Stemming']:
                        #print(s)
                        if s['Soort'] in ['Voor', 'Tegen']:
                            info[z['Nummer']]['Stem_'+s['ActorFractie']]=(s['Soort'])
                            info[z['Nummer']]['Aantal_stemmen_'+str(s['ActorFractie'])]=(s['FractieGrootte'])
                        #info[z['Nummer']]['Vergissing_'+str(s['ActorFractie'])]=(s['Vergissing']) niet nodig, want Stem is altijd de gecorrigeerde stem (dus niet de vergissing)
                    #print('stemmingenverwerkt')
                    break
            for a in z['Agendapunt']:
                info[z['Nummer']]['AgendapuntOnderwerp']=a['Onderwerp']

            for d in z['Document']:
                if d['Soort'][:5]=='Motie':
                    #print('\n','doc',d)
                    info[z['Nummer']]['doc_Id']=d['Id']
                    info[z['Nummer']]['Volgnummer']=d['Volgnummer']
                    info[z['Nummer']]['Datum']=d['Datum']
                    if INCLUDE_PDF:
                        #print('doing pdf work')
                        retrieve_and_save_pdf(d['Id'])
                        info[z['Nummer']]['Text']=pdf_to_text(path+d['Id']+'.pdf')
            
            info[z['Nummer']]['Indiener_persoon'] = []
            info[z['Nummer']]['Medeindiener_persoon'] = []
            for za in z['ZaakActor']:
                #print('\n','za',za)
                if za['Relatie'] == 'Indiener' and za['ActorFractie']:
                    info[z['Nummer']]['Indiener_persoon'].append(za['ActorNaam'])
                    info[z['Nummer']]['Indiener_'+za['ActorFractie']]=1
                
                if za['Relatie'] == 'Medeindiener' and za['ActorFractie']:
                    info[z['Nummer']]['Medeindiener_persoon'].append(za['ActorNaam'])
                    info[z['Nummer']]['Medeindiener_'+za['ActorFractie']]=1
    except Exception as e: 
        print(e)
        pprint.pprint(z) 
        traceback.print_exc() 
        print(s)
        sys.exit()


### Get all moties

# main routine to query api
FIRST_ONLY = False
INCLUDE_PDF = False
START = 0
MAXIMUM = 100000
path = 'pdf/' # where to store pdfs of motions
skip = 0
info = {}
print ('started')
data = query_API(str(START))
count = data['@odata.count']
# add_API_to_dict(data)
skip = START +  250

if not FIRST_ONLY:
    while skip < count + 250 and skip < MAXIMUM:
        print('query for ',skip)
        data = query_API(str(skip))
        add_API_to_dict(data)
        skip +=250
print('finished')

import pickle
with open('moties_unprocessed.pickle', 'wb') as handle:
    pickle.dump(info, handle, protocol=pickle.HIGHEST_PROTOCOL)

## Download PDFs

import pickle
file = open("moties_unprocessed.pickle","rb")
info = pickle.load(file)

download_errors = set()
no_doc_id = set()

# takes a long time
for i, (k,v) in enumerate(info.items()):
    if i % 100 == 0: print(i)
    doc_id = v.get('doc_Id', None)
    if doc_id:
        try:
            retrieve_and_save_pdf(doc_id)
            info[k]['Text'] = pdf_to_text2(path + doc_id +'.pdf')
        except Exception as e:
            print('download or pdf_to_text error')
            traceback.print_exc() 
            download_errors.add((k,doc_id))
    else:
        no_doc_id.add(k)
df_text = pd.DataFrame(info).T

# Some pdf download may have been corrupted. By running this function, we try to redownload the pdf and update the text. Unfortunatly it seems the pdfs are simply corrupted
no_text_idx = df_text[(df_text['Text']=='') | (df_text['Text'].isna())].index
print(len(no_text_idx), 'no texts found')
texts = []
for index in no_text_idx:
    doc_Id = df_text.loc[index,'doc_Id']
    print(doc_Id, type(doc_Id))
    if not pd.isna(doc_Id):
        retrieve_and_save_pdf(doc_Id, overwrite=True)
        t = pdf_to_text(path + doc_Id + '.pdf')
        if t != '':
            df_text.loc[index,'Text'] = t
            texts.append(t)
print(len(texts), 'new pdfs found')

import pickle
with open('df_text_before_preprocessing', 'wb') as handle:
    pickle.dump(df_text, handle, protocol=pickle.HIGHEST_PROTOCOL)

## Make dataframe

import pickle
file = open("df_text_before_preprocessing","rb")
df = pickle.load(file)
print(len(df))
column_list = df.columns.values # current columns

df.iloc[0]['Text']

### Preprocessing

def preprocessing_motie(text,doc_id,i): # remove start and end of document to only include relevant text
    try:
        if text and doc_id and text != 'nan':
            try:
                regex = re.findall(r"\d{4}(.*?)Nr.*gehoord de beraadslaging(.*)",text,re.DOTALL)
                regex = (' '.join(regex[0]))
            except Exception:
                regex = re.findall(r"\d{4}(.*?)",text,re.DOTALL)
                regex = (' '.join(regex[0]))
            return regex
        else:
            if not doc_id:
                doc_id.append(i)
            elif not text:
                no_text.append(i)
            return ""
    except Exception:
        print(traceback.print_exc())
        error_processing.append(i)
        return ""


# df['Text'] = df['Text'].astype(str)
# df['Text'] = df['Text'].str.replace('\xad', '')
# df['Text'] = df['Text'].str.replace('\n', '')
# df['Text'] = df['Text'].str.replace('\r', '')
# df['Text'] = df['Text'].str.replace('-', '')

# forget about text preprocessing, do that in later notebook with LDA
import numpy as np
no_text = []
error_processing = []
doc_id = []
# df['Text_processed'] = np.vectorize(preprocessing_motie)(df['Text'],df['doc_Id'],df.index)
len(no_text),len(error_processing),len(doc_id)

#verwijder hoofdelijke stemmingen
print (len(df), 'removing hoofdelijke stemmingen')
df.drop(df[df['StemmingsSoort'] == 'Hoofdelijk'].index, inplace=True)
print (len(df), '\n')

#verwijder moties zonder stemming
print (len(df), 'removing moties without vote')
print(df['BesluitSoort'].value_counts())
df = df[((df['BesluitSoort'] == 'Stemmen - aangenomen') | (df['BesluitSoort'] == 'Stemmen - verworpen'))]
print (len(df), '\n')

#recode besluitsoort naar -1 en 1
print(df['BesluitSoort'].value_counts())
df['BesluitSoort']=df['BesluitSoort'].replace({'Stemmen - verworpen':'0','Stemmen - aangenomen':'1'})
df['BesluitSoort']=pd.to_numeric(df['BesluitSoort'])
print(df['BesluitSoort'].value_counts(), '\n')

#recode voor en tegen naar 1 en -1
stem_column = [c for c in column_list if 'Stem_' in c]
print(df['Stem_50PLUS'].value_counts())
df[stem_column]=df[stem_column].replace({'Tegen':'-1','Voor':'1','Niet deelgenomen':np.nan})
print(df['Stem_50PLUS'].value_counts(), '\n')

#cast to datetime and sort old to new
df['Datum'] = pd.to_datetime(df['Datum'].str[:10]) # cut off the time & timezone info, we only need date
# df['Datum'] = pd.to_datetime(df['Datum'], utc=True)
df['GestartOp'] = pd.to_datetime(df['GestartOp'].str[:10])
# df['GestartOp'] = pd.to_datetime(df['GestartOp'], utc=True)
# df['Datum'] = df['Datum'].dt.tz_convert(None)
# df['GestartOp'] = df['GestartOp'].dt.tz_convert(None)
df.sort_values('Datum',inplace=True)

column_list = df.columns.values
indiener_column = [c for c in column_list if ('Indiener') in c and c != 'Indiener_persoon']
# make a column with the indienende partij
ind = pd.melt(df, value_vars=indiener_column, ignore_index=False)
ind = ind[ind['value'] == 1]
ind = ind.rename(columns={'variable':'Indienende_partij'})
ind['Indienende_partij'] = ind['Indienende_partij'].str[9:]
df['Indiener_persoon'] = df['Indiener_persoon'].map(lambda x: x[0] if x else '')
df = df.join(ind['Indienende_partij'], how='left')
df['Indienende_persoon_partij'] = df['Indiener_persoon'] + ', ' + df['Indienende_partij']

#bereken voor en tegenstemmen
aantal_stemmen_column = [c for c in column_list if 'Aantal' in c]

for i in range(len(stem_column)):
    df[stem_column[i]]=pd.to_numeric(df[stem_column[i]])
    df[aantal_stemmen_column[i]]=pd.to_numeric(df[aantal_stemmen_column[i]])

res = pd.DataFrame(np.multiply(df[stem_column].fillna(0).values, df[aantal_stemmen_column].fillna(0).values))
voor = res[res > 0].sum(axis=1)
tegen = abs(res[res < 0].sum(axis=1))
df['Voor'], df['Tegen'] = voor.values,tegen.values
df['Delta'] = abs(df['Voor']- df['Tegen'])
df['Sum'] = (df['Voor']+ df['Tegen'])


df = df.replace(-1,0)

df['Jaar'] = df['Datum'].dt.year

# only 26 moties for first year, so let's chop
df = df[(df['Jaar'] > 2008) & (df['Jaar'] < 2025)]

from collections import namedtuple
import datetime
Cabinet = namedtuple('Cabinet', ['name','start', 'end', 'demissionair','coalition'])
Chamber = namedtuple('Chamber', ['name','start', 'end', 'numyears'])

chamber_input = [
    "Balkenende IV,01 01 2009,16 06 2010",# mark begin 2008 is not correct (should be 30 11 2006) but for dataset purpose it it
    "Rutte I,17 06 2010,19 09 2012",
    "Rutte II,20 09 2012,22 03 2017",
    "Rutte III,23 03 2017,30 03 2021", # mark end 2020 is not correct but for dataset purpose it it
    "Rutte IV,31 03 2021,31 12 2024"] 

chambers = {}
for c in chamber_input:
    c = c.split(',')
    name = c[0]
    start = list(map(int, c[1].split()))
    startdate = datetime.datetime(start[2], start[1], start[0])
    end = list(map(int, c[2].split()))
    enddate = datetime.datetime(end[2], end[1], end[0])
    numyears = (enddate - startdate).days/365
    chambers[name] = Chamber(name, startdate, enddate, numyears)


cabinets_input = [
    # "Balkenende III,07 07 2006,22 02 2007,CDA VVD",
    # "Balkenende IV,22 02 2007,14 10 2010,CDA PvdA ChristenUnie",
    "Balkenende IV,18 09 2008,13 10 2010,20 02 2010,CDA PvdA ChristenUnie",
    "Rutte I,14 10 2010,4 11 2012,23 04 2012,VVD CDA PVV",
    "Rutte II,5 11 2012,25 10 2017,14 03 2017,VVD PvdA",
    "Rutte III,26 10 2017,09 01 2022,15 01 2022,CDA VVD D66 ChristenUnie",
    "Rutte IV,10 01 2022, 01 01 2024,07 07 2023,VVD D66 CDA ChristenUnie"]

cabinets = {}
for c in cabinets_input:
    c = c.split(',')
    name = c[0]
    start = list(map(int, c[1].split()))
    startdate = datetime.datetime(start[2], start[1], start[0])
    end = list(map(int, c[2].split()))
    enddate = datetime.datetime(end[2], end[1], end[0])
    demissionair = list(map(int, c[3].split()))
    demissionair_date = datetime.datetime(demissionair[2], demissionair[1], demissionair[0])
    coalition = tuple(c[4].split())
    cabinets[name] = Cabinet(name, startdate, enddate, demissionair_date, coalition)
cabinets, chambers

import datetime
def get_period(entities, date):
    # date = datetime.datetime.strptime(date_string[:10], '%Y-%m-%d')
    # date = date.replace(tzinfo=None)
    for c in entities.values():
        if ((c.start <= date) or (date < datetime.datetime(2009,1,1))) and c.end >= date:
            return c.name
    else:
        return 'Error'


def get_demissionair(cabinets, date):
    # date = datetime.datetime.strptime(date_string[:10], '%Y-%m-%d')
    c = get_period(cabinets, date)
    # date = date.replace(tzinfo=None)

    return False if date < cabinets[c].demissionair else True
    
df['Kabinet'] = [get_period(cabinets, d) for d in df['Datum']]
df['Demissionair'] = [get_demissionair(cabinets, d) for d in df['Datum']]
df['Kamer'] = [get_period(chambers, d) for d in df['Datum']]
df['Kamer'].value_counts(), df['Kabinet'].value_counts()

INCLUDE_PDF = True

# sort columns
column_list = df.columns.values # current columns
column_order = ['Titel','AgendapuntOnderwerp','Onderwerp','Datum','GestartOp','Vergaderjaar','Jaar','Indiener_persoon', 'Indienende_partij','Indienende_persoon_partij','Kamer', 'Kabinet','Demissionair','StemmingsSoort','BesluitSoort','BesluitTekst','doc_Id','Volgnummer', 'Voor', 'Tegen', 'Delta', 'Sum', 'Text'] # desired columns

not_in_column_order = set(column_list) - set(column_order)
 # don't throw away columns
column_order += sorted(not_in_column_order)
df = df[column_order]


cnt = 0
def process_text(text):
    idx = text.find('De Kamer')
    global cnt
    if idx != -1:
        cnt += 1
    return text if idx == -1 else text[idx:]

def remove_kst(text):
    idx = text.find('kst-')
    return text if idx == -1 else text[:idx]

def find_end(text):
    idx = text.find(', en gaat over tot')
    return text if idx == -1 else text[:idx] + '.'

df['Text'] = df['Text'].fillna('')
df['ClippedText'] = df['Text']
# if you remove the begin part you are removing info about the subject and ministery
# df['ClippedText'] = df['Text'].apply(process_text)
# df['ClippedText'].apply(process_text)
df['ClippedText'] = df['ClippedText'].str.replace('\n', ' ')
df['ClippedText'] = df['ClippedText'].str.replace(r'- [^e]', '', regex=True)
df['ClippedText'] = df['ClippedText'].apply(remove_kst)
df['ClippedText'] = df['ClippedText'].apply(find_end)


#df = df[df['Onderwerp'].str.contains('klimaat')]
import pickle
with open('moties_processed_df.pickle', 'wb') as handle:
    pickle.dump(df.copy(), handle, protocol=pickle.HIGHEST_PROTOCOL)

### Making data ready for ChatGPT

import pickle
import pandas as pd
file = open("moties_processed_df.pickle","rb")
df = pickle.load(file)
df = df[df['Kamer']=='Rutte IV']
df['Text'].str.len().sum()


df.Indienende_partij.value_counts()


import string
partij2idx = {p: 'PARTIJ_' + string.ascii_uppercase[idx] for idx, p in enumerate(df.Indienende_partij.unique())}
df['Prompt'] = 'De volgende motie is ingediende door ' + df['Indienende_partij'].apply(lambda x: partij2idx[x]) + ' \n' + df['ClippedText']


df.drop(df[df['ClippedText']==""].index, inplace=True)
with open('moties_for_llm.pickle', 'wb') as handle:
    pickle.dump(df, handle, protocol=pickle.HIGHEST_PROTOCOL)
