==> 3_top2vec_analysis.ipynb <==
from pathlib import Path
from top2vec import Top2Vec
import pickle
import numpy as np

from gensim.models import Phrases
from gensim.models.phrases import Phraser
from gensim.utils import simple_preprocess
from gensim.parsing.preprocessing import strip_tags

DATA_DIR = Path('./data')
def load(filename):
    f = open(DATA_DIR/filename,"rb")
    return pickle.load(f)
    
def save(data, filename):
    with open(DATA_DIR/filename, 'wb') as handle:
        pickle.dump(data, handle, protocol=pickle.HIGHEST_PROTOCOL)

def add_voor_tegen_index(df):
    print(len(df))
    stem_column = [c for c in df.columns if 'Stem_' in c and c != 'Stem_persoon']
    stem_column_adj = [c[5:] for c in df.columns if 'Stem_' in c and c != 'Stem_persoon']
    stem_array = df[stem_column].values.tolist()
    assert len(stem_array[0]) == len(stem_column)
    voor = [[stem_column_adj[i] for i, stem in enumerate(motie) if stem == 1] for motie in stem_array]
    tegen = [[stem_column_adj[i] for i, stem in enumerate(motie) if stem == 0] for motie in stem_array]
    df['Partijen_Voor'] = voor
    df['Partijen_Tegen'] = tegen
    df['Index']=list(range(len(df)))
    return df

def get_df(only_last_kamer):
    file = open("moties_processed_df.pickle","rb")
    df = pickle.load(file)
    print('before removal empty texts',len(df))

    # remove moties without text
    mask = (df['Text']=='') | (df['Text'].isna())
    df = df.loc[~mask]
    print('after removal empty texts',len(df))
    if only_last_kamer:
        df = df[df['Kabinet']==last_kamer]
        print('after selecting latest cabinet', len(df))
    df = add_voor_tegen_index(df)
    df['BesluitTekst'] = df['BesluitTekst'].str.replace('.','')
    return df, df['ClippedText'].values

## Model Training

only_last_kamer = False
last_kamer = 'Rutte IV'

if only_last_kamer:
    df, documents = get_df(True)
else:
    df, documents = get_df(False)
    
# got the code for making the bigram part from https://lppier.github.io/
sentence_stream = [simple_preprocess(strip_tags(doc), deacc=True) for doc in documents]
bigram = Phrases(sentence_stream, min_count=30)
bigram_phraser = Phraser(bigram)

indieners = {indiener[-1].lower() for indiener in df['Indiener_persoon'].str.split() if indiener}
years = {word for doc in sentence_stream for word in doc if word.startswith('x') and not word.startswith('xin')}
manual_stopwords = {'faber', 'lacin', 'kroger', 'iv', 'beschikt', 'die', 'vaststelling', 'lid','vi', 'viii', 'iii', 'iv', 'ii', 'i', 'kamer', 'regering tevens', 'regering', 'gehoord','beraadslaging', 'overweegt', 'overwegende', "gehoord","regering","gesproken","horen","mondkapjes","spreken","regeringen","gesprekken","commissiedebat","discussies","discussie","hoort","regeringsbeleid","besprekingen","overheden","overheid","overheids","overheidsbeleid","gesprek","zeggen","dialoog","geluid","uitspreekt","bureaucratische","spreekt","openbare","audiovisuele","gehoor","bestuursakkoord","stil","minuten","parlement","parlementen","officieel","bureaucratie","democratie","officiele","batenanalyse","parlementair","begrotings","governance","democratische","budgetrecht","democratisch","kabinetsreactie","belastingplan","wetsvoorstellen","wetsvoorstel","thema","dicht","risicoanalyse","bestuursrecht","bestuursorganen","european","nederlandsche","nederlanden","eurogroep","nederlander","nederland","holland","hogescholen","no","kinder","po","nee","verzoek","verzoeken","gevraagde","gevraagd","aangevraagd","verzoekt","claims","euro","eurocommissaris","vergunning","instemming","educatieve","eisen","voorstel","aanvrager","afgewezen","toepassing","voorstelt","aanvragen","vereist","werkloosheid","energietoeslag","ongewenste","aanvraag","behoefte","for","voorgesteld","weigert","automatisch","wachten","binnenkort","graag","gat","dji","mond","non","imago","toch","makkelijk","gestuurd","onmogelijk","nu","pas","gemakkelijk","co","eenvoudig","dende","natuurlijk","vervolg","uitzetten","zeker","opriep","budgettaire","fiscale","gefinancierde","financieel","gefinancierd","duizenden","financiele","financierings","ongekend","whw","ja","nertsen","zo","inderdaad","gevaarlijk","budget","begrotingen","budgetten","begroting","overgebleven","uitgaven","kostenbesparing","subsidies","overtreden","overtreding","onverminderd","overtredingen","weren","obstakels","afgeschaft","thans","rulings","uitsluiting","subsidieren","over","reduceren","overschrijden","onwenselijk","hinder","to","preventief","overschrijding","beperking","overtuiging","voorkomen","bestedingen","promoten","sanctie","verhinderen","nr","je","leeftijdsgrens","kleinere","wel","nibud","niet","novi","verzamelt" ,"fatsoen" ,"imams" ,"imam" ,"minuut" ,"fantastisch" ,"keurig" ,"terechtgekomen" ,"bik" ,"verschijnt" ,"belem" ,"bel" ,"vermo" ,"zins" ,"ima" ,"zoe" ,"sappen" ,"verdien" ,"kundig" ,"straks" ,"viteit" ,"mondig" ,"mooi" ,"letten" ,"combi" ,"geweldig" ,"sak" ,"bene" ,"faso" ,"wijde" ,"hoor" ,"verrekend" ,"prachtige" ,"fao" ,"teneinde" ,"capten" ,"openlijk" ,"openhouden" ,"tege" ,"simpele" ,"tak" ,"simpel" ,"korrels" ,"hww" ,"geleden" ,"meteen" ,"volstaat" ,"lees" ,"weet" ,"ver" ,"eenvoudiger" ,"wetswijzigingen" ,"inmiddels" ,"hbo" ,"ziet" ,"makkelijker" ,"zitten" ,"snelle" ,"drukke" ,"snel" ,"iack" ,"gekeken" ,"kijkt" ,"vindt" ,"dak" ,"btw" ,"lijken" ,"ots" ,"kijken" ,"gemakkelijker" ,"geacht" ,"eenvoudige" ,"zodra" ,"abro" ,"verzamelen" ,"gelukt" ,"goed" ,"mooie" ,"tenzij" ,"hakken" ,"grondwettelijke" ,"korte" ,"spoedig" ,"nh" ,"laat" ,"stank" ,"daaronder" ,"kortst" ,"klaar" ,"krap","uitstekende" ,"ben" ,"begrijpen","daarover","leggen","opties","hierover","toegevoegd","aangetoond","dat","drie","veilig","daarmee","kelijk","beschreven","weten","mogen","hen","gegaan" }
stopwords = indieners | years | manual_stopwords


def bigram_stopword_preprocess(doc):
    sentence_stream = simple_preprocess(strip_tags(doc), deacc=True)
    sentence_stream = [word for word in sentence_stream if word not in stopwords]
    return sentence_stream
    # return bigram_phraser[sentence_stream]

df['Kamer'].value_counts()

# https://github.com/scikit-learn-contrib/hdbscan/issues/607

model = Top2Vec(documents, speed='deep-learn', embedding_model='distiluse-base-multilingual-cased',workers=8, min_count=10, tokenizer=bigram_stopword_preprocess, ngram_vocab=False)
model.get_num_topics()
if only_last_kamer:
    model.save("data/doc2vec_deep_bigram_enhanced_stopwords_last_kamer")
else:
    model.save("data/doc2vec_deep_bigram_enhanced_stopwords_all")

## Reduce topic number to something more manageable

# For the manual analysis it's necessary to reduce the amount of topics, otherwise you will have 50 or so topics to investigate

df, documents = get_df(only_last_kamer)
if only_last_kamer:
    model = Top2Vec.load("data/doc2vec_deep_bigram_enhanced_stopwords_last_kamer")
else:
    model = Top2Vec.load("data/doc2vec_deep_bigram_enhanced_stopwords_all")

topic_sizes, topic_nums = model.get_topic_sizes()
print(f'{len(topic_nums)} topics. In total there are {sum(topic_sizes)} documents. These are the amount of documents per topic:\n{topic_sizes}')

# helper to check what happens if you reduce topics
def get_reduced_topics(num_topics):
    print(f'performing reduction to {num_topics} topics')
    reduced_topics = model.hierarchical_topic_reduction(num_topics)
    topic_words, word_scores, topic_nums = model.get_topics(reduced=True)
    reduced_topics = tuple(tuple(sorted(t)) for t in reduced_topics)
    return reduced_topics, topic_words

def find_diff(reduced1, reduced2, reverse=False):
    # find the topics that where merged and return their index
    if not reverse:
        changed = set(reduced1) - set(reduced2)
        return [index for index, topic in enumerate(reduced1) if topic in changed]
    else:
        changed = set(reduced2) - set(reduced1)
        return [index for index, topic in enumerate(reduced2) if topic in changed]


def print_merge(large, small, num_words=50):
    print(f'\ninspecting difference from {len(large[0])} to {len(small[0])} topics')
    print('old topics')
    for i in find_diff(large[0],small[0]):
        print(large[1][i][:num_words])
    print('new topic')
    for i in find_diff(large[0],small[0], reverse=True):
        print(small[1][i][:num_words])

def find_optimal_num_topics(minn, maxx):
    start = minn
    stop = maxx
    reductions = {i: get_reduced_topics(i) for i in range(start, stop, -1)}
    for i in range(start, stop +1,-1):
        print(i)
        print_merge(reductions[i], reductions[i-1])
find_optimal_num_topics(17,12)

# Make a choice to how many topics to reduce
num_topics = 17
reduced_topics = model.hierarchical_topic_reduction(num_topics)


topic_words, word_scores, topic_nums = model.get_topics(reduced=True)
'" ,"'.join(t for t in topic_words[0])


topic_words, word_scores, topic_nums = model.get_topics(reduced=True)
for i, t in enumerate(topic_words):
    print(i, t)

topic_words, word_scores, topic_nums = model.get_topics(reduced=False)
# for t in topic_words:
#     print(t)

# get some more stopwords to filter out
from collections import defaultdict
sums = defaultdict(int)

for t, s in zip(topic_words, word_scores):
    for word, score in zip(t,s):
        sums[word] += score + 1
    
scores = [(-score, word)  for word, score in sums.items()]
scores.sort()
'","'.join(s for w,s in scores if '_' not in s)

topic_words, word_scores, topic_scores, topic_nums = model.search_topics(keywords= ['politie'] , num_topics=4, reduced=True)


topics = {
  0: 'Algemene zaken',  
  1: 'Zorg',  
  2: 'Gemeente',  
  3: 'Onderwijs',  
  4: 'Werk & Pensioen',  
  5: 'Milieu & Klimaat',  
  6: 'Landbouw & Dierenwelzijn',  
  7: 'Rechtsstaat',  
  8: 'Immigratie',  
  9: 'Vervoer',  
  10: 'Overig',  
  11: 'Financien',  
  12: 'Justitie',  
  13: 'Kabinetsbeleid',  
  14: 'Jeugd',  
  15: 'Energie',
  16: 'Digitalisering'
  }
# topics = {
#   0: 'Onderwijs',  
#   1: 'Buitenlandse zaken',  
#   2: 'Algemene zaken',  
#   3: 'Natuur & gaswinning',  
#   4: 'Landbouw & dierenwelzijn',  
#   5: 'Zorg',  
#   6: 'Sociale zaken',  
#   7: 'Justitie',  
#   8: 'Pensioenstelsel',  
#   9: 'Europese Unie',  
#   10: 'Klimaat & energie',  
#   11: 'Milieu & regelgeving',  
#   12: 'Zorg',  
#   13: 'Openbaar vervoer',  
#   14: 'Financiele sector',  
#   15: 'Wonen'
#   }
topic_words, word_scores, topic_nums = model.get_topics()
# topics = {i:' ,'.join(topic_words[i][:3]) for i in range(len(topic_words))}
doc_ids = list(range(len(documents)))
topic_nums, topic_score, topic_words, word_scores = model.get_documents_topics(doc_ids,reduced=True)
topic_names = [topics[t] for t in topic_nums]
assert len(topic_nums) == len(df)
df['Topic'] = topic_names
topic_nums, topic_score, topic_words, word_scores = model.get_documents_topics(doc_ids,reduced=False)
df['Topic_initial'] = topic_nums
df['Topic_score'] = topic_score
# df.sort_values(['Topic_initial', 'Topic_score'], ascending=False, inplace=True)


topics

df['Topic'].unique()

## Optional add climate deepdive

reduced_topics

topic_sizes, topic_nums = model.get_topic_sizes(reduced=False)
topic_words, word_scores, topic_nums = model.get_topics(reduced=False)
climate_idx = [topic_nums for t in topic_nums if t in reduced_topics[5]]
climate_idx


topic_sizes, topic_nums = model.get_topic_sizes(reduced=False)
topic_words, word_scores, topic_nums = model.get_topics(reduced=False)
climate_idx = [topic_nums for t in topic_nums if t in reduced_topics[10]]
climate_subtopics = {128: 'Afhankelijkheid fossiele brandstoffen',
 165: 'CO2 reductie',
 5: 'Voldoen aan Parijs',
 141: 'Electriciteit',
 126: 'Groningen',
 205: 'Zonnepanelen',
 96: 'Energierekening betalen',
 236: 'Biomassa',
 105: 'Kolencentrales',
 239: 'Windturbines - overlast',
 29: 'Windturbines - subsidie'}

df['Klimaat'] = df.loc[df['Topic']=='Klimaat & energie', 'Topic_initial'].map(climate_subtopics)

save(df, 'df_including_topics_full.pickle')
save(model, 'doc2vec_deep_bigram_enhanced_stopwords_rutteIV_reduced')

# Prepare slimmed down versions for production

df = load("df_including_topics_full.pickle")
# df, documents = get_df(only_last_kamer)
# df['BesluitTekst'] = df['BesluitTekst'].str.replace('.','')

print(len(df))

if only_last_kamer:
    model = Top2Vec.load("data/doc2vec_deep_bigram_enhanced_stopwords_last_kamer")
else:
    model = Top2Vec.load("data/doc2vec_deep_bigram_enhanced_stopwords_all")
    
doc_ids = list(range(len(df)))
# don't sort the df before this operation
topic_nums, topic_score, topic_words, word_scores = model.get_documents_topics(doc_ids,reduced=False)
topics = [', '.join(topic_words[i][:3]) for i in range(len(topic_words))]

df['Topic_initial'] = topics
df['Topic_score'] = topic_score
df.sort_values(['Topic_initial', 'Topic_score'], ascending=False, inplace=True)
# important do this only after all row filters have been set
df = df[df['Kamer']== last_kamer]
df.index = df['Index']
print(len(df))
stem_column = [c for c in df.columns if 'Stem_' in c and c != 'Stem_persoon']
required_cols = ['Kamer', 'Jaar','Indienende_partij', 'BesluitSoort','BesluitTekst','Topic_initial', 'Topic_score','Indienende_persoon_partij','Partijen_Voor', 'Partijen_Tegen', 'Text']

# streamlit has problems with category type: 
# https://github.com/streamlit/streamlit/issues/47
for col in stem_column + required_cols[:-4]:
    df[col] = df[col].astype('category')

df = df[stem_column + required_cols]
save(df, "df_production.pickle")


def delete_documents(self, doc_ids):
    """
    Delete documents from current model.
    Warning: If document ids were not used in original model, deleting
    documents will change the indexes and therefore doc_ids.
    The documents will be deleted from the current model without changing
    existing document, word and topic vectors. Topic sizes will be updated.
    If deleting a large quantity of documents relative to the current model
    size a new model should be trained for best results.
    Parameters
    ----------
    doc_ids: List of str, int
        A unique value per document that is used for referring to documents
        in search results.
    """
    # make sure documents exist
    self._validate_doc_ids(doc_ids, doc_ids_neg=[])

    # update index
    if self.documents_indexed:
        # delete doc_ids from index
        index_ids = [self.doc_id2index_id(doc_id) for doc_id in doc_ids]
        for index_id in index_ids:
            self.document_index.mark_deleted(index_id)
        # update index_id and doc_ids
        for doc_id in doc_ids:
            self.doc_id2index_id.pop(doc_id)
        for index_id in index_ids:
            self.index_id2doc_id.pop(index_id)

    # get document indexes from ids
    doc_indexes = self._get_document_indexes(doc_ids)

    # delete documents
    if self.documents is not None:
        self.documents = np.delete(self.documents, doc_indexes, 0)

    # delete document ids
    if self.document_ids is not None:
        for doc_id in doc_ids:
            self.doc_id2index.pop(doc_id)
        keys = list(self.doc_id2index.keys())
        self.document_ids = np.array(keys)
        values = list(range(0, len(self.doc_id2index.values())))
        self.doc_id2index = dict(zip(keys, values))

    # delete document vectors
    self._set_document_vectors(np.delete(self._get_document_vectors(norm=False), doc_indexes, 0))

    if self.embedding_model == 'doc2vec':
        num_docs = len(doc_indexes)
        self.model.docvecs.count -= num_docs
        self.model.docvecs.max_rawint -= num_docs
        self.model.docvecs.vectors_docs_norm = None
        self.model.docvecs.init_sims()

    # update topics
    # self._unassign_documents_from_topic(doc_indexes, hierarchy=False)

    if self.hierarchy is not None:
        self._unassign_documents_from_topic(doc_indexes, hierarchy=True)
model = Top2Vec.load("data/doc2vec_deep_bigram_enhanced_stopwords_rutteIV_reduced")

# model.delete_documents = delete_documents
df = load("df_including_topics_full.pickle")


topic_sizes, topic_nums = model.get_topic_sizes()
print(f'{len(topic_nums)} topics. In total there are {sum(topic_sizes)} documents. These are the amount of documents per topic:\n{topic_sizes}')

topic_words, word_scores, topic_nums = model.get_topics(reduced=True)
    
for num, size, word in zip(topic_nums, topic_sizes, topic_words):
    print(num, word[:3], size)
print(sum(topic_sizes))

model.delete_documents(list(range(len(df))))
model.delete_documents = 1
model.save("data/doc2vec_production")

## Check app

df = load("df_production.pickle")
model = Top2Vec.load("data/doc2vec_production")
topic_words, word_scores, topic_scores, topic_nums = model.search_topics(keywords= ['politie'] , num_topics=4, reduced=True)
topic_nums

topic_sizes, topic_nums = model.get_topic_sizes()
print(f'{len(topic_nums)} topics. In total there are {sum(topic_sizes)} documents. These are the amount of documents per topic:\n{topic_sizes}')

topic_words, word_scores, topic_scores, topic_nums = model.search_topics(keywords= ['politie'] , num_topics=4, reduced=True)
topic_nums

topic_words, word_scores, topic_nums = model.get_topics(num_topics=47)


topic_words[45]

df = load("df_production.pickle")
len(df)

df[(df['Topic_initial']=='verduurzamen') & (df['Kamer']== last_kamer)]['Text'].iloc[0]

df = load("df_including_topics_full.pickle")
print(len(df))
df['BesluitTekst'].value_counts()
