==> 6_llm_principle_analysis copy.ipynb <==
import urllib.request
import json
import numpy as np
import pandas as pd
import altair
import pprint
import os, subprocess, sys,codecs, locale
import re
import traceback
from collections import namedtuple
import datetime
import pickle
import pandas as pd
import openai
from pprint import pprint
from aocutils.datascience import save, load

openai.organization = ""
openai.api_base = "https://openrouter.ai/api/v1"
openroutertoken = open('.tokens/test2.txt','r').read()
openai.api_key = openroutertoken
OPENROUTER_REFERRER = "https://github.com/jvanelteren/kamermoties"

# openai.Model.list()

from openai import OpenAI
from os import getenv

# gets API Key from environment variable OPENAI_API_KEY
client = OpenAI(
  base_url="https://openrouter.ai/api/v1",
  api_key=openroutertoken,
)
def AI(text):
    completion = client.chat.completions.create(

    model="openai/gpt-4-1106-preview",
    messages=[
                {"role": "system", "content": "You are a great programmer"},
                {"role": "user", "content": text},
            ])
    return completion.choices[0].message.content

AI('Given a list of multiple line equations in general form with coeficients a,b,c,d, write a function that determines a line that intersects all other lines')

from openai import OpenAI
from os import getenv

# gets API Key from environment variable OPENAI_API_KEY
client = OpenAI(
  base_url="https://openrouter.ai/api/v1",
  api_key=openroutertoken,
)
def AI(text):
    completion = client.chat.completions.create(

    model="openai/gpt-4-1106-preview",
    messages=[
                {"role": "system", "content": "You are the greatest botanist alive. You know everything of plants, shrubs and trees. You give expert advice on what plants do well in certain gardens. Your answers are short. Users will prompt you with certain questions what grows well in their garden."},
                {"role": "user", "content": text},
            ])
    return completion.choices[0].message.content

text = 'I have a garden where I want to grow plants that are great for insects and butterflies with nice flowers. The soil type is slightly acidic and dense with sand underneath. There are some oaks growing in the garden, so it is a bit foresty. We live in the Netherlands and close to heath, which explains the acidic soil. The oaks take away most of the sunlight and use a lot of the rain. What plants would do well in my garden? As mentioned, they should be able to grow well in shade and on sandy soil, without too much water'
res = AI(text)
pprint.pprint(res)

text = 'I have a garden where I want to grow edible plants. The soil type is slightly acidic and dense with sand underneath. There are some oaks growing in the garden, so it is a bit foresty. We live in the Netherlands and close to heath, which explains the acidic soil. The oaks take away most of the sunlight and use a lot of the rain. What plants would do well in my garden? As mentioned, they should be able to grow well in shade and on sandy soil, without too much water'
res = AI(text)
pprint.pprint(res)

text = 'I have a garden where I want to grow plants as a natural fence. The soil type is slightly acidic and dense with sand underneath. There are some oaks growing in the garden, so it is a bit foresty. We live in the Netherlands and close to heath, which explains the acidic soil. The oaks take away most of the sunlight and use a lot of the rain. What plants would do well in my garden? Rhodondendron and Cherry Laurel already do well, but I want something else. I want plants that grow quite densely and fast, to a minimum of 1-2 meters high, wintergreen, which are good for biodiversity, possibly with some flowers. And as mentioned, they should be able to grow well in shade and on sandy soil, without too much water'
res = AI(text)
pprint.pprint(res)

pprint.pprint(res)

pprint.pprint(res)

import pprint
pprint.pprint(res)

pprint.pprint(res)
